+++ MercadoLibre/Modules/Product/ProductViewController.swift
@@ -12,6 +12,10 @@ class ProductViewController: UIViewController, Storyboarded {
     // MARK: - IBOutlets -
 
     @IBOutlet weak var titleLabel: UILabel!
+    @IBOutlet weak var activityIndicator: UIActivityIndicatorView!
+    @IBOutlet weak var mainImageView: UIImageView!
+    @IBOutlet weak var priceLabel: UILabel!
+    @IBOutlet weak var descriptionLabel: UILabel!
 
     // MARK: - Public properties -
 
@@ -25,6 +29,7 @@ class ProductViewController: UIViewController, Storyboarded {
         super.viewDidLoad()
 
         // Do any additional setup after loading the view.
+        viewModel?.loadProductById()
     }
 
 }
@@ -32,15 +37,16 @@ class ProductViewController: UIViewController, Storyboarded {
 extension ProductViewController: ProductView {
 
     func startLoading() {
-
+        activityIndicator.startAnimating()
     }
 
     func finishLoading() {
-
+        activityIndicator.stopAnimating()
     }
 
     func setData(with product: Product) {
-
+        titleLabel.text = product.title
+        priceLabel.text = "\(product.price)"
     }
 
     func setEmptyView() {